ImportTradeDoublerHotels (not finished)
git-svn-id: https://subversion.assembla.com/svn/importproducts/trunk@98 c09a0911-d79b-4797-ba2f-bc3321b6aa18

+++ ImportProducts/Hotel.cs
@@ -16,8 +16,8 @@ namespace ImportProducts
     {
         public Nullable<int> LocationId { get; set; }
         public Nullable<int> Rooms { get; set; }
-        public Nullable<double> Star { get; set; }
-        public Nullable<double> CustomerRating { get; set; }
+        public Nullable<decimal> Star { get; set; }
+        public Nullable<decimal> CustomerRating { get; set; }
         public string Address { get; set; }
         public string CurrencyCode { get; set; }
